Domotic/services/requirements.txt
paho-mqtt
psycopg2
dotenv